pyproject.toml
[tool.poetry]
name = "kc-moe"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kc_moe"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
konlpy = "^0.6.0"
python-mecab-ko = "^1.2.9"
dask = {extras = ["distributed"], version = "^2022.12.1"}
ipykernel = "^6.19.2"
fastparquet = "^2022.12.0"
wikipedia-api = "^0.5.8"
tensorflow = "^2.11.0"
seqio = {git = "https://github.com/google/seqio"}
jax = {version = "^0.4.1", extras = ["tpu"], optional = true}
tensorflow-datasets = "4.7"
gcsfs = "^2022.11.0"
gin-config = "^0.5.0"
nervaluate = "^0.1.8"
t5x = {git = "https://github.com/google-research/t5x.git", branch="main", extras = ["tpu"], optional = true}
t5 = {git = "https://github.com/google-research/text-to-text-transfer-transformer.git", branch="main", optional = true}
flaxformer = {git = "https://github.com/google/flaxformer.git", branch="main", optional = true} 
more-itertools = "^9.0.0"

[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/libtpu_releases.html"
default = false
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
